+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Changelog
 
+## 2.1.0
+- Add basic unit tests with THREE.js mocking.
+- Refactor userData as a separate module.
+
 ## 2.0.0
 - Make graphics, and other input parameters, dynamic.
 
 ## 1.0.0
 - Convert client-side code part of a monolith to a reusable module.
+
+# Future features to develop
+- Validate input of skybox.
+- Validate input of userData.
+- Add more tests.